Separated Pcap-VMS from the rest of the simh source tree.
This is being done since the plan going forward (in the primary descrip.mms and the Visual Studio Projects) is to local pcap components (VMS and Windows) one level up from the simh directory.

The PCAP-VMS components are presumed (by this procedure) to be located
in a directory at the same level as the directory containing the
simh source files.  For example, if these exist here:

   []descrip.mms
   []scp.c
   etc.

Then the following should exist:
   [-.PCAP-VMS]BUILD_ALL.COM
   [-.PCAP-VMS.PCAP-VCI]
   [-.PCAP-VMS.PCAPVCM]
   etc.