Remove ECLIPSE from the VAX build since it requires 64b and move the presumed location of PCAP-VMS up one directory level.
This is being done so that it is not entangled with the rest of
the simh sources and is at the same level presumed for the Windows
pcap components presumed in the Visual Studio Projects.

The PCAP-VMS components are presumed (by this procedure) to be located
in a directory at the same level as the directory containing the
simh source files.  For example, if these exist here:

  []descrip.mms
  []scp.c
  etc.

Then the following should exist:

  [-.PCAP-VMS]BUILD_ALL.COM
  [-.PCAP-VMS.PCAP-VCI]
  [-.PCAP-VMS.PCAPVCM]
  etc.