LP20 emulation was overwriting existing files
When an LP20 is attached to an existing file, the file is opened in
update mode.  Since the initial file position is zero, any existing
output is over-written.  This is not intuitive.

This edit makes the LP20 seek to EOF on attach.

N.B. Other devices may have cousin issues.