Improve handling of off-line/unattached condition
CD11 wasn't starting off-line with the hopper empty, even though no file
is attached.
Hardware reset was preserving off-line and hopper empty, but not Error.

Thus:
If a file wasn't attached, OFFLINE would not be set initially.
If a file was attached, the OS would read cards before the RESET switch
(set cr reset) was pressed.

This doesn't match the hardware.

Initialize CDST to off-line with an empty hopper.
Always set ERR with OFFLINE instead of relying on CSR read to do it, as
it's also checked internally.

CR11 also wasn't initialized to off-line and had similar problems.
When writing crs, ERR was cleared even if the unit remained off-line.
ERR should be set if the unit is off-line.