RH11/TM0x issues: attach/detach
Tape on-line interrupts weren't happening when a tape was attached or
replace on the TH11/TM0x emulation.

The OS never saw a MOL transition in the case of replacement, where the
detach is implicit.

This patch defers reporting MOL for 100 msec; faster than a human can
change tapes, but long enough for an OS to see the 1->0->1 transition of
MOL.

Further, checks for UNIT_ATT should be for MOL.

This makes GALAXY automatic tape recognition work.