QVSS: Fix to handle when cursor is near the right edge of the screen and only partially visible.
Prior fix for cursor tracking allowed this to now be observed.

The description of the prior change should have been:

The status bits need to be set in the upper byte of the DUART receiver buffer. This is not mentioned in the DEC manual but I found a Phillips datasheet for the 2681 chip, which has a small footnote about it.