SDS: Allow console bootstrap from RAD if connected to W channel
Normally the RAD is connected to the E channel and there is no bootstrap capability (because of hardwired WIM instruction). However, for completeness, add code to allow booting from the RAD if it is connected to the W channel. Add RAD_CHAN definition to define which controller the RAD is connected to.

Improve parentheses in bootstrap code that allows an optional word count (while also permitting the canonical bootstrap sequence).