SAGE and sigma fixes.
SAGE: Fixes a loop index's starting value.
cache_line has a size of CACHE_SIZE (16), so starting at that index will be outside the bounds of the array.

sigma_cpu: Fix an always true case
This was the equivalent of "if (vec >= 0x52 || vec <= 0x55)".

sigma_lp: Fix a fopen leak