AltairZ80: Fix head stepping to only occur on 1 to 0 transitions.
The source file to be updated provides emulation of the North Star DD
floppy disk controller. The change steps the head only when the step bit
in the order byte transitions from 1 to 0. This is the way the actual
hardware functions. Previously, the emulation would step any time an
order was issued and the step bit was 1. This behavior caused some
original software (e.g., Lifeboat CP/M 2.2) to end up on an incorrect
track.