Add support for North Star single density controller to AltairZ80
The first disk controller made by North Star was a single density controller (MDSA).
This was followed by the double density controller (MDSAD) that is already supported
in SIMH. This update adds support for the single density controller as device
MDSA. Since the controllers are not software compatible, this update allows
running of older software designed for the MDSA controller.